安装结束后解压源码删除  2

+++ plugins/php/versions/common/sodium.sh
@@ -50,9 +50,11 @@ Install_lib()
 		php_lib=$sourcePath/php_lib
 		mkdir -p $php_lib
 		if [ ! -d $php_lib/${LIBNAME}-${LIBV} ];then
-			wget -O $php_lib/${LIBNAME}-${LIBV}.tgz http://pecl.php.net/get/lib${LIBNAME}-${LIBV}.tgz
+			if [ ! -f $php_lib/${LIBNAME}-${LIBV}.tgz ];then
+				wget -O $php_lib/${LIBNAME}-${LIBV}.tgz http://pecl.php.net/get/lib${LIBNAME}-${LIBV}.tgz
+			fi
 			cd $php_lib && tar xvf ${LIBNAME}-${LIBV}.tgz
-		fi 
+		fi
 		cd $php_lib/lib${LIBNAME}-${LIBV}
 
 		$serverPath/php/$version/bin/phpize
